Delete duplicate Login page
