+++ frontend/src/pages/Login.jsx
@@ -1,54 +1,71 @@
-import React from 'react';
-import './Login.css';
-import logo from '../images/cas-logo.png';
-import {Row, Col, Container, Form, Button} from "react-bootstrap";
+import { useState } from "react";
+import "./Login.css";
+import logo from "../images/cas-logo.png";
+import { Button, Container, Form } from "react-bootstrap";
 
 const Login = () => {
-    return <>
-    <Row className="mt-5">
-        <Col lg={1}></Col>
-        <Col lg={5} md={6} sm={12} className="p-5 m-auto">
-            <Container className="p-5 mt-5 shadow-sm rounded-lg bg-white text-dark bg-opacity-50">
-                <h4>Privacy Notice</h4>
-                <p>privacy notice details will be placed here<br/>
-                Lorem ipsum dolor sit amet, consectetur adipiscing elit. 
-                Nulla quis est metus. Vestibulum a libero vitae est eleifend convallis vel ut justo. 
-                Pellentesque in augue molestie, sollicitudin elit accumsan, luctus purus. Donec ut fringilla nisl. 
-                Praesent vitae tincidunt justo. Cras ullamcorper, 
-                erat quis bibendum euismod, nisl velit tincidunt arcu, accumsan efficitur ipsum quam eu lectus. 
-                Sed id dignissim enim. Praesent dolor sapien, bibendum eget auctor vitae, placerat id elit. 
-                Etiam ultrices ultrices sapien, a iaculis dui sagittis nec. Phasellus mi est, accumsan vel euismod at, semper quis sapien. 
-                Nulla arcu nisi, rhoncus at commodo in, porttitor nec arcu. Donec nisl sapien, tempus at dapibus sed, hendrerit eget turpis. 
-                Ut lobortis semper elementum. Nulla facilisi.</p>
-            </Container>
-        </Col>
+  const [email, setEmail] = useState("");
+  const [password, setPassword] = useState("");
 
-        <Col>
-            <Container className="mt-5 p-5">
-            <img src={logo} width="150" height="150" className="logo"/>
-            <br/>
-            <h2>GWA Verifier</h2>
-            <br/>
-                <Form className="login-form">
-                    <Form.Group controlId="formEmail" className="w-50">
-                        <Form.Label className="input-label">Email</Form.Label>
-                        <Form.Control type="email" title="Enter email"/>
-                    </Form.Group>
-                    <br/>
-                    <Form.Group controlId="formPassword" className="w-50">
-                        <Form.Label className="input-label">Password</Form.Label>
-                        <Form.Control type="password" title="Enter password"/>
-                    </Form.Group>
+  function login(e) {
+    e.preventDefault();
 
-                    <br/>
-                    <Button variant="outline-light w-50" type="submit" className="login-button">
-                        Login
-                    </Button>
-                </Form>
-            </Container>
-        </Col>
-    </Row>
-    </>;
-  };
+    const credentials = {
+      email: email,
+      password: password,
+    };
+
+    fetch("/login", {
+      method: "POST",
+      headers: {
+        "Content-Type": "application/json",
+      },
+      body: JSON.stringify(credentials),
+    })
+      .then((response) => response.json())
+      .then((body) => {
+        if (!body.success) {
+          alert("Failed to log in");
+        } else {
+          alert("Successfully logged in");
+          console.log(credentials);
+        }
+      });
+  }
+
+  return (
+    <>
+      <div className="leftHalf"></div>
+      <div className="rightHalf">
+        <Container className="mt-5 p-5">
+          <img src={logo} width="150" height="150" className="logo" />
+          <h2>GWA Verifier</h2>
+          <br />
+          <Form onSubmit={login} className="login-form">
+            <Form.Group controlId="formEmail" className="w-50">
+              <Form.Label className="input-label">Email</Form.Label>
+              <Form.Control type="email" title="Enter email" value={email} onChange={(e) => setEmail(e.target.value)} />
+            </Form.Group>
+            <br />
+            <Form.Group controlId="formPassword" className="w-50">
+              <Form.Label className="input-label">Password</Form.Label>
+              <Form.Control
+                type="password"
+                title="Enter password"
+                value={password}
+                onChange={(e) => setPassword(e.target.value)}
+              />
+            </Form.Group>
+
+            <br />
+            <Button variant="outline-light w-50" type="submit" className="login-button">
+              Login
+            </Button>
+          </Form>
+        </Container>
+      </div>
+    </>
+  );
+};
 
 export default Login;
