--- frontend/src/pages/LoginPage.css
@@ -1,54 +1,0 @@
-html, body {
-    color: maroon;
-}
-
-.leftHalf {
-    background: url(../images/kk-tower.jpg);
-    width: 50%;
-    position: absolute;
-    left: 0px;
-    height: 100%;
-    background-size: cover;
-    background-repeat: no-repeat;
-    background-position: center center;
-}
-
-.rightHalf {
-    width: 50%;
-    position: absolute;
-    right: 0px;
-    height: 100%;
-}
-
-.logo{
-    display: block;
-    margin-left: auto;
-    margin-right: auto;
-    width: 100;
-    height: 100;
-}
-
-h2{
-    text-align: center;
-    color: maroon;
-    margin: 10px;
-}
-
-.login-form{
-    display: grid;
-    place-items: center;
-}
-
-.form-control:focus {
-    border-color: maroon;
-    box-shadow: inset 0 1px 1px rgba(0, 0, 0, 0.075), 0 0 8px rgba(255, 0, 0, 0.6);
-}
-
-.input-label{
-    color: maroon;
-    width: 100%;
-}
-
-.login-button{
-    background-color:maroon;
-}
